feat(hrneo): reused NDMS catalogue
