+++ frontend/src/lib/components/hrneo/HrNeoEditModal.svelte
@@ -5,10 +5,10 @@
 		PolicyGlobalInterface,
 		RoutingTunnel,
 	} from '$lib/types';
-	import { presetCatalog } from '$lib/stores/presets';
 	import type { CatalogPreset } from '$lib/types';
-	import { Modal, Button, IconButton, Dropdown, type DropdownOption } from '$lib/components/ui';
-	import { CatalogPresetRow, IconPickerModal, ServiceIcon } from '$lib/components/dnsroutes';
+	import { Modal, Button, Dropdown, type DropdownOption } from '$lib/components/ui';
+	import { IconPickerModal, ServiceIcon } from '$lib/components/dnsroutes';
+	import { presetDnsEntryCount, splitPresetDnsEntries } from '$lib/utils/catalog-preset';
 	import { InterfaceList } from '$lib/components/accesspolicy';
 	import HrNeoGeoTagPicker from './HrNeoGeoTagPicker.svelte';
 	import { buildRoutingTunnelDropdownOptions } from '$lib/utils/routingTunnelOptions';
@@ -31,6 +31,8 @@
 		maxelem: number;
 		saving: boolean;
 		initialTarget?: { kind: 'interface' | 'policy'; name: string };
+		initialPreset?: CatalogPreset | null;
+		onpickcatalog?: () => void;
 		onsave: (payload: Partial<DnsRoute>) => void;
 		onclose: () => void;
 	}
@@ -46,12 +48,11 @@
 		maxelem,
 		saving,
 		initialTarget,
+		initialPreset = null,
+		onpickcatalog,
 		onsave,
 		onclose,
 	}: Props = $props();
-
-	// Only presets with inline domains can be used — HR has no subscriptions.
-	let usablePresets = $derived($presetCatalog.filter((p) => (p.engines.dns?.domains?.length ?? 0) > 0));
 
 	let name = $state('');
 	let iconUrl = $state<string | undefined>(undefined);
@@ -65,7 +66,6 @@
 	let newPolicyName = $state('');
 	let newPolicyIfaces = $state<AccessPolicyInterface[]>([]);
 
-	let presetPickerOpen = $state(false);
 	let selectedPreset = $state<CatalogPreset | null>(null);
 	let geositePickerOpen = $state(false);
 	let geoipPickerOpen = $state(false);
@@ -99,7 +99,6 @@
 		selectedPreset = null;
 		geositePickerOpen = false;
 		geoipPickerOpen = false;
-		presetPickerOpen = false;
 		iconPickerOpen = false;
 		if (rule) {
 			name = rule.name;
@@ -196,26 +195,30 @@
 			initialExistingPolicyName = existingPolicyName;
 			initialNewPolicyName = newPolicyName;
 			initialNewPolicyIfaces = [];
+			if (initialPreset) {
+				applyPresetFields(initialPreset);
+			}
 		}
 	});
 
-	function applyPreset(p: CatalogPreset) {
-		selectedPreset = p;
-		const entries = p.engines.dns?.domains ?? [];
-		const domainLines: string[] = [];
-		const cidrLines: string[] = [];
-		for (const e of entries) {
-			if (e.startsWith('geoip:') || /^[\d.:a-fA-F]+\/\d+$/.test(e)) cidrLines.push(e);
-			else domainLines.push(e);
+	let lastAppliedPresetId = $state<string | null>(null);
+
+	$effect(() => {
+		if (!open) {
+			lastAppliedPresetId = null;
+			return;
 		}
+		if (!initialPreset || initialPreset.id === lastAppliedPresetId) return;
+		lastAppliedPresetId = initialPreset.id;
+		applyPresetFields(initialPreset);
+	});
+
+	function applyPresetFields(p: CatalogPreset) {
+		selectedPreset = p;
+		const { domainLines, cidrLines } = splitPresetDnsEntries(p);
 		domainsText = domainLines.join('\n');
 		cidrText = cidrLines.join('\n');
 		if (!name.trim()) name = p.name;
-		presetPickerOpen = false;
-	}
-
-	function clearPreset() {
-		selectedPreset = null;
 	}
 
 	function appendLine(which: 'domains' | 'cidr', token: string) {
@@ -282,6 +285,12 @@
 	let newPolicyNameValidationError = $derived(hrPolicyNameError(newPolicyName));
 
 	const interfaceTunnelOpts = $derived(buildRoutingTunnelDropdownOptions(tunnels));
+
+	let hasDomainOrCidrContent = $derived(
+		splitLines(domainsText).length > 0 || splitLines(cidrText).length > 0,
+	);
+
+	let showCatalogEntry = $derived(!!onpickcatalog && !hasDomainOrCidrContent);
 
 	let canSave = $derived.by(() => {
 		if (!name.trim()) return false;
@@ -370,35 +379,39 @@
 </script>
 
 <Modal {open} {title} size="lg" {onclose} hasUnsavedChanges={() => isDirty}>
-	<!-- Preset bar -->
-	<div class="preset-bar">
-		<div class="preset-bar-left">
+	{#if showCatalogEntry}
+		<div class="catalog-entry">
 			{#if selectedPreset}
-				<ServiceIcon name={selectedPreset.name} iconSlug={selectedPreset.iconSlug} size={24} />
-				<div class="preset-bar-info">
-					<div class="preset-bar-name">{selectedPreset.name}</div>
-					<div class="preset-bar-meta">{selectedPreset.engines.dns?.domains?.length ?? 0} записей</div>
-				</div>
-				<IconButton ariaLabel="Очистить пресет" onclick={clearPreset}>×</IconButton>
+				<button type="button" class="catalog-picked" onclick={onpickcatalog}>
+					<ServiceIcon
+						name={selectedPreset.name}
+						iconSlug={selectedPreset.iconSlug}
+						size={28}
+					/>
+					<div class="catalog-entry-info">
+						<div class="catalog-entry-name">{selectedPreset.name}</div>
+						<div class="catalog-entry-meta">
+							{presetDnsEntryCount(selectedPreset)} записей из каталога
+						</div>
+					</div>
+					<span class="catalog-change">Сменить</span>
+				</button>
 			{:else}
-				<span class="preset-bar-label">Пресет не выбран</span>
+				<button type="button" class="catalog-cta" onclick={onpickcatalog}>
+					<span class="catalog-cta-icon" aria-hidden="true">
+						<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.75">
+							<rect x="3" y="3" width="7" height="7" rx="1" />
+							<rect x="14" y="3" width="7" height="7" rx="1" />
+							<rect x="3" y="14" width="7" height="7" rx="1" />
+							<rect x="14" y="14" width="7" height="7" rx="1" />
+						</svg>
+					</span>
+					<span class="catalog-cta-text">
+						<span class="catalog-cta-title">Заполнить из каталога</span>
+						<span class="catalog-cta-hint">Выберите сервис — подставим домены и CIDR</span>
+					</span>
+				</button>
 			{/if}
-		</div>
-		<Button variant="secondary" size="sm" onclick={() => (presetPickerOpen = !presetPickerOpen)}>
-			{presetPickerOpen ? 'Скрыть каталог' : 'Выбрать из каталога'}
-		</Button>
-	</div>
-
-	{#if presetPickerOpen}
-		<div class="preset-catalog">
-			{#each usablePresets as p (p.id)}
-				<CatalogPresetRow
-					name={p.name}
-					iconSlug={p.iconSlug}
-					meta={`${p.engines.dns?.domains?.length ?? 0} записей`}
-					onclick={() => applyPreset(p)}
-				/>
-			{/each}
 		</div>
 	{/if}
 
@@ -666,50 +679,94 @@
 		text-overflow: ellipsis;
 	}
 
-	.preset-bar {
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-		gap: 12px;
-		padding: 10px 12px;
-		background: var(--color-bg-secondary);
-		border: 1px solid var(--color-border);
-		border-radius: 8px;
+	.catalog-entry {
 		margin-bottom: 14px;
 	}
-	.preset-bar-left {
+
+	.catalog-cta,
+	.catalog-picked {
 		display: flex;
 		align-items: center;
-		gap: 10px;
-		min-width: 0;
+		gap: 12px;
+		width: 100%;
+		padding: 12px 14px;
+		border-radius: 10px;
+		border: 1px solid var(--color-border);
+		background: var(--color-bg-primary);
+		cursor: pointer;
+		font-family: inherit;
+		text-align: left;
+		transition:
+			border-color 0.15s,
+			background 0.15s,
+			box-shadow 0.15s;
 	}
-	.preset-bar-info {
+
+	.catalog-cta {
+		border-color: color-mix(in srgb, var(--color-accent) 45%, var(--color-border));
+		background: color-mix(in srgb, var(--color-accent) 8%, var(--color-bg-primary));
+	}
+
+	.catalog-cta:hover,
+	.catalog-picked:hover {
+		border-color: var(--color-accent);
+		box-shadow: 0 0 0 1px color-mix(in srgb, var(--color-accent) 25%, transparent);
+	}
+
+	.catalog-cta-icon {
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		width: 40px;
+		height: 40px;
+		border-radius: 8px;
+		background: color-mix(in srgb, var(--color-accent) 14%, transparent);
+		color: var(--color-accent);
+		flex-shrink: 0;
+	}
+
+	.catalog-cta-text {
 		display: flex;
 		flex-direction: column;
+		gap: 2px;
 		min-width: 0;
 	}
-	.preset-bar-name {
+
+	.catalog-cta-title {
+		font-size: 0.875rem;
 		font-weight: 600;
 		color: var(--color-text-primary);
 	}
-	.preset-bar-meta {
+
+	.catalog-cta-hint {
+		font-size: 0.75rem;
+		color: var(--color-text-muted);
+	}
+
+	.catalog-picked {
+		background: var(--color-bg-secondary);
+	}
+
+	.catalog-entry-info {
+		flex: 1;
+		min-width: 0;
+	}
+
+	.catalog-entry-name {
+		font-weight: 600;
+		color: var(--color-text-primary);
+	}
+
+	.catalog-entry-meta {
 		color: var(--color-text-muted);
 		font-size: 0.75rem;
 	}
-	.preset-bar-label {
-		color: var(--color-text-muted);
-		font-size: 0.875rem;
-	}
 
-	.preset-catalog {
-		display: flex;
-		flex-direction: column;
-		background: var(--color-bg-secondary);
-		border: 1px solid var(--color-border);
-		border-radius: 8px;
-		margin-bottom: 14px;
-		max-height: 280px;
-		overflow-y: auto;
+	.catalog-change {
+		font-size: 0.75rem;
+		font-weight: 600;
+		color: var(--color-accent);
+		flex-shrink: 0;
 	}
 
 	.form-section {
